requirements.txt
# Core dependencies - Python 3.13 compatible
fastapi==0.115.5
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pydantic==2.10.3
pydantic-settings==2.7.0
python-magic==0.4.27